[Update] Removed all stuff not needed for cel shader
