--- libs/cel-shader/.gitignore
@@ -1,12 +1,0 @@
-CMakeCache.txt
-CMakeFiles
-Makefile
-cmake_install.cmake
-install_manifest.txt
-build/
-
-*.dll
-*.lib
-*.so
-*.a
-*.o
