--- libs/cel-shader/README.md
@@ -1,33 +1,0 @@
-Cel Shader
-==================
-
-A simple GLSL based cel shading example. I originally wrote this in 2007.
-
-Required libraries: OpenGL, GLU, (free)glut, GLEW
-
-* glut/freeglut can be downloaded from here: http://freeglut.sourceforge.net/
-* GLEW can be downloaded from here: http://glew.sourceforge.net/
-
-This program uses GLSL and requires an OpenGL implementation, and hardware, that is compatibale with the OpenGL 2.0 standard and higher.
-
-The model used in the third scene was obtained from: 
-http://www.katorlegaz.com/3d_models/ 
-and is licensed under a Creative Commons Attribution 3.0 United States License and is Copyright © 2003-2012 Andrew Kator & Jennifer Legaz.
-It was exported from Blender with a customized raw binary exporter.
-
-Keys
-----
-* _Spacebar or most other unused keys_: go to the next scene
-* _up arrow_: adjust the camera's pitch -
-* _down arrow_: adjust the camera's pitch +
-* _left arrow_: rotate the camera around the scene in the clockwise direction
-* _right arrow_: rotate the camera around the scene in the counter clockwise direction
-* _+_: move the camera closer to the object(s)
-* _-_: move the camera away from the object(s)
-* _Esc_: quits the program
-
-Author
-------
-Morn&#xe9; Chamberlain
-
-http://github.com/mchamberlain
