--- libs/cel-shader/src/MiscGL.cpp
@@ -1,100 +1,0 @@
-// Copyright (c) 2012, ME Chamberlain
-// All rights reserved.
-// 
-// Redistribution and use in source and binary forms, with or without modification,
-// are permitted provided that the following conditions are met:
-// 
-// 	- Redistributions of source code must retain the above copyright notice, this
-// 	  list of conditions and the following disclaimer.
-// 	- Redistributions in binary form must reproduce the above copyright notice,
-// 	  this list of conditions and the following disclaimer in the documentation 
-// 	  and/or other materials provided with the distribution.
-// 
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
-// EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
-// OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
-// SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
-// INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
-// TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR
-// BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY
-// WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-#include <string>
-#include <cmath>
-#include <GL/glew.h>
-#include <GL/glut.h>
-
-#include "MiscGL.h"
-
-void MiscGL::setWindow(GLfloat x0, GLfloat y0, GLfloat x1, GLfloat y1, float boundary) {
-	float deltaWidth;
-	float deltaHeight;
-
-	if (boundary != 0.0f) {
-		deltaWidth = boundary * fabs(x1 - x0) / 2.0f;
-		deltaHeight = boundary * fabs(y1 - y0) / 2.0f;
-
-		if (deltaWidth == 0) {
-			deltaWidth = deltaHeight;
-		}
-		else if (deltaHeight == 0) {
-			deltaHeight = deltaWidth;
-		}
-
-		x0 -= deltaWidth;
-		y0 -= deltaHeight;
-		x1 += deltaWidth;
-		y1 += deltaHeight;
-	}
-
-	glMatrixMode(GL_PROJECTION);
-	glLoadIdentity();
-	gluOrtho2D(x0, x1, y0, y1);
-}
-
-void MiscGL::setViewport(GLint x, GLint y, GLfloat width, GLfloat height, float boundary) {
-	float deltaWidth;
-	float deltaHeight;
-
-	if (boundary != 0.0f) {
-		deltaWidth = boundary * width;
-		deltaHeight = boundary * height;
-
-		if (deltaWidth == 0) {
-			deltaWidth = deltaHeight;
-		}
-		else if (deltaHeight == 0) {
-			deltaHeight = deltaWidth;
-		}
-
-		x -= (GLint) (deltaWidth / 2.0f);
-		y -= (GLint) (deltaHeight / 2.0f);
-		width += deltaWidth;
-		height += deltaHeight;
-	}
-
-	glViewport(x, y, (GLsizei) width, (GLsizei) height);
-}
-
-// Copyright (c) 2012, ME Chamberlain
-// All rights reserved.
-// 
-// Redistribution and use in source and binary forms, with or without modification,
-// are permitted provided that the following conditions are met:
-// 
-// 	- Redistributions of source code must retain the above copyright notice, this
-// 	  list of conditions and the following disclaimer.
-// 	- Redistributions in binary form must reproduce the above copyright notice,
-// 	  this list of conditions and the following disclaimer in the documentation 
-// 	  and/or other materials provided with the distribution.
-// 
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
-// EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
-// OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
-// SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
-// INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
-// TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR
-// BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY
-// WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
