+++ libs/celshader/.gitignore
@@ -1,0 +1,12 @@
+CMakeCache.txt
+CMakeFiles
+Makefile
+cmake_install.cmake
+install_manifest.txt
+build/
+
+*.dll
+*.lib
+*.so
+*.a
+*.o
